Move Dockerfile generation to a separate package
I intend to use the `pkg/docker` package as wrappers around Docker
commands.
